refactor xmate3
