+++ mani_skill2/agents/robots/xmate3/__init__.py
@@ -1,2 +1,1 @@
-from .configs import Xmate3RobotiqDefaultConfig
 from .xmate3 import Xmate3Robotiq
